// demo/src/payload-types.ts
export interface Document {
  id: number;
  title: string;
  content: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  metaTitle?: string | null;
  metaDescription?: string | null;
  canonicalUrl?: string | null;
  genericText1?: string | null;
  genericText2?: string | null;
  slug: string;
  publishedAt?: string | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}